requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
readability-lxml==0.8.1
python-dotenv==1.0.1
google-genai
groq
tavily-python
